# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RFP Demo Node API",
    "version": "1.0.0",
    "description": "API documentation for RFP Demo Node"
  },
  "servers": [
    {
      "url": "https://rfp-api-service-c8adb8a8h2gmbcf8.canadacentral-01.azurewebsites.net"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for MCP server integration"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/api/send-email": {
      "post": {
        "tags": [
          "Email"
        ],
        "summary": "Send generated proposal via email",
        "description": "Sends the generated proposal to a specified email address using Azure Communication Services Email. The email includes formatted HTML content with the proposal text.\n",
        "operationId": "sendProposalEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "subject",
                  "proposalText"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Recipient email address",
                    "format": "email"
                  },
                  "subject": {
                    "type": "string",
                    "description": "Email subject"
                  },
                  "proposalText": {
                    "type": "string",
                    "description": "Generated proposal text content"
                  },
                  "companyName": {
                    "type": "string",
                    "description": "Company name (optional)"
                  },
                  "savedAs": {
                    "type": "string",
                    "description": "Blob file name reference (optional)"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical email request",
                  "value": {
                    "to": "client@example.com",
                    "subject": "Proposal for AI RFP Automation",
                    "proposalText": "Executive Summary\n\nWe are pleased to submit..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "to": {
                      "type": "string"
                    },
                    "messageId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or invalid email format"
          },
          "500": {
            "description": "Email sending failed"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/generate": {
      "post": {
        "tags": [
          "Proposal Generation"
        ],
        "summary": "Generate a full proposal from extracted RFP text",
        "description": "This endpoint orchestrates the full proposal generation workflow using extracted RFP text.\nIt parses deadlines and criteria, checks compliance, extracts structured requirements, performs optional company research, retrieves similar past proposals, and generates a complete proposal document.\nThis route exists as the primary entry point for automated proposal generation. An MCP server can infer that this endpoint coordinates multiple AI services.\n",
        "operationId": "generateProposal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "extractedText"
                ],
                "properties": {
                  "extractedText": {
                    "type": "string",
                    "description": "Full extracted plain text of the RFP"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical request",
                  "value": {
                    "extractedText": "The client seeks a cloud migration partner..."
                  }
                },
                "edgeCase": {
                  "summary": "Edge case: minimal text",
                  "value": {
                    "extractedText": "RFP for cafeteria services."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proposal generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "companyName": {
                      "type": "string",
                      "nullable": true
                    },
                    "rfpData": {
                      "type": "object"
                    },
                    "compliance": {
                      "type": "object"
                    },
                    "requirementsJson": {
                      "type": "object"
                    },
                    "companyResearch": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "retrievedEvidence": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "proposal": {
                      "type": "string"
                    },
                    "savedAs": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid extractedText, or compliance score below threshold",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "compliance_score": {
                      "type": "number"
                    },
                    "missing_sections": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "present_sections": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error during proposal generation"
          }
        },
        "x-sideEffects": [
          "Calls AI services for extraction and generation",
          "Calls semantic search",
          "Writes proposal output to Blob Storage"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/parser/parse": {
      "post": {
        "tags": [
          "RFP Parsing"
        ],
        "summary": "Parse extracted RFP text into structured metadata",
        "description": "Parses extracted RFP text and returns structured metadata such as deadlines and evaluation criteria.\nThis route exists to convert unstructured text into machine-readable data.\n",
        "operationId": "parseRFPText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "extractedText"
                ],
                "properties": {
                  "extractedText": {
                    "type": "string",
                    "description": "Full extracted plain text of the RFP"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical RFP text",
                  "value": {
                    "extractedText": "Request for Proposal\n1. Introduction\nABC Corporation invites proposals..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed RFP metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deadline": {
                      "type": "string"
                    },
                    "criteria": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid extractedText"
          },
          "500": {
            "description": "Parser failure"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/parser/parse-from-blob": {
      "post": {
        "tags": [
          "RFP Parsing"
        ],
        "summary": "Parse RFP metadata from a blob-stored file",
        "description": "Downloads an RFP file from blob storage, extracts its text, and parses structured metadata.\n",
        "operationId": "parseRFPFromBlob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "blobName"
                ],
                "properties": {
                  "blobName": {
                    "type": "string",
                    "description": "Name of the blob file in incoming container"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical blob reference",
                  "value": {
                    "blobName": "123e4567-RFP.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed metadata from blob file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deadline": {
                      "type": "string"
                    },
                    "criteria": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing blobName"
          },
          "500": {
            "description": "Blob or parser failure"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/parser/compliance": {
      "post": {
        "tags": [
          "RFP Parsing"
        ],
        "summary": "Check RFP compliance with required sections",
        "description": "Checks whether an RFP text contains required proposal sections.\nUsed for validation and quality control before generation.\n",
        "operationId": "checkRFPCompliance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "extractedText"
                ],
                "properties": {
                  "extractedText": {
                    "type": "string",
                    "description": "Full extracted plain text of the RFP"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical RFP text",
                  "value": {
                    "extractedText": "Request for Proposal\n1. Introduction\nABC Corporation invites proposals..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compliance check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "compliance_score": {
                      "type": "string",
                      "description": "Percentage score (e.g., \"37.5%\")"
                    },
                    "present_sections": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of sections found in the RFP"
                    },
                    "missing_sections": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of required sections not found"
                    }
                  }
                },
                "examples": {
                  "incomplete": {
                    "summary": "Incomplete RFP result",
                    "value": {
                      "compliance_score": "37.5%",
                      "present_sections": [
                        "Executive Summary or Introduction",
                        "Scope of Work"
                      ],
                      "missing_sections": [
                        "Timeline or Schedule",
                        "Evaluation Criteria"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing extractedText"
          },
          "500": {
            "description": "Compliance check failure"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/download-pdf": {
      "post": {
        "tags": [
          "PDF Generation"
        ],
        "summary": "Generate and download proposal as PDF",
        "description": "Converts the generated proposal text into a professionally formatted PDF document with branding, metadata, and structured content sections.\n",
        "operationId": "downloadProposalPDF",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "proposal"
                ],
                "properties": {
                  "proposal": {
                    "type": "string",
                    "description": "Generated proposal text content"
                  },
                  "companyName": {
                    "type": "string",
                    "description": "Company name (optional)"
                  },
                  "rfpData": {
                    "type": "object",
                    "description": "RFP metadata (optional)"
                  },
                  "requirementsJson": {
                    "type": "object",
                    "description": "Requirements data (optional)"
                  },
                  "savedAs": {
                    "type": "string",
                    "description": "Blob file name reference for filename (optional)"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical PDF generation request",
                  "value": {
                    "proposal": "Executive Summary\n\nWe are pleased to submit our proposal...",
                    "companyName": "ABC Corporation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF file generated successfully",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Missing required proposal text"
          },
          "500": {
            "description": "PDF generation failed"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/proposals/retrieve": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Retrieve similar past proposals",
        "description": "Uses semantic similarity search to find past proposals relevant to the provided RFP text and optional industry.\nThis route exists to support proposal research and reuse.\n",
        "operationId": "retrieveSimilarProposals",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "extractedText"
                ],
                "properties": {
                  "extractedText": {
                    "type": "string",
                    "description": "Full extracted plain text of the RFP"
                  },
                  "industry": {
                    "type": "string",
                    "description": "Optional industry filter (e.g., \"healthcare\", \"finance\")"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical retrieval request",
                  "value": {
                    "extractedText": "Request for Proposal for cloud migration services...",
                    "industry": "technology"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of similar proposals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "industry": {
                      "type": "string",
                      "nullable": true
                    },
                    "proposals": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "similarityScore": {
                            "type": "number"
                          },
                          "content": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing extractedText"
          },
          "500": {
            "description": "Search service failure"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/proposals/retrieve-from-blob": {
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Retrieve similar proposals from blob-stored RFP",
        "description": "Downloads an RFP file from blob storage, extracts its text, and retrieves similar past proposals using semantic search.\n",
        "operationId": "retrieveSimilarProposalsFromBlob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "blobName"
                ],
                "properties": {
                  "blobName": {
                    "type": "string",
                    "description": "Name of the blob file in incoming container"
                  },
                  "industry": {
                    "type": "string",
                    "description": "Optional industry filter (e.g., \"healthcare\", \"finance\")"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical blob retrieval request",
                  "value": {
                    "blobName": "123e4567-RFP.pdf",
                    "industry": "healthcare"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of similar proposals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "blobName": {
                      "type": "string"
                    },
                    "industry": {
                      "type": "string",
                      "nullable": true
                    },
                    "proposals": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing blobName"
          },
          "500": {
            "description": "Blob download or search failure"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/upload": {
      "post": {
        "tags": [
          "RFP Upload"
        ],
        "summary": "Upload an RFP file and extract its text content",
        "description": "This endpoint accepts an RFP (Request for Proposal) document as a file upload (PDF, DOCX, etc.), stores it in Azure Blob Storage, and extracts its text content for downstream processing.\nInternally, a unique file ID is generated, the file is uploaded to blob storage, and a parser service extracts plain text.\nThis route exists to provide a standardized ingestion point for RFP documents, decoupling file handling from proposal generation. An MCP server can infer that this endpoint is for document ingestion, not proposal creation.\n",
        "operationId": "uploadRFP",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "rfp": {
                    "type": "string",
                    "format": "binary",
                    "description": "RFP file to upload (PDF, DOCX, etc.)"
                  }
                }
              },
              "examples": {
                "basic": {
                  "summary": "Typical file upload",
                  "value": {
                    "rfp": "(binary)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully uploaded and extracted RFP text",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "fileId": {
                      "type": "string"
                    },
                    "blobName": {
                      "type": "string"
                    },
                    "extractedTextPreview": {
                      "type": "string"
                    },
                    "extractedText": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "summary": "Successful upload",
                    "value": {
                      "message": "RFP uploaded successfully",
                      "fileId": "123e4567-e89b-12d3-a456-426614174000",
                      "blobName": "123e4567-RFP.pdf",
                      "extractedTextPreview": "The client requests...",
                      "extractedText": "The client requests a cloud migration..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No file uploaded or invalid input"
          },
          "500": {
            "description": "Internal server error during upload or extraction"
          }
        },
        "x-sideEffects": [
          "Uploads file to Azure Blob Storage",
          "Extracts text from the uploaded file"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "tags": []
}